+++ src/load.h
@@ -1,5 +1,7 @@
+#include <stddef.h>
+
 #pragma once
 
 void initialize_load();
-void update_load(unsigned long);
+void update_load(size_t);
 void show_load();
